fix: PostgreSQL-compatible SQL types in Flyway migrations; MODE=PostgreSQL in H2 test URLs
- DOUBLE → DOUBLE PRECISION in V1000, V1001, V1002
- Added MODE=PostgreSQL to H2 JDBC URL in test application.properties

Closes #66

+++ runtime/src/main/resources/db/migration/V1002__ledger_supplement.sql
@@ -49,7 +49,7 @@ CREATE TABLE ledger_supplement_compliance (
     -- GDPR Art.22 / EU AI Act Art.12
     decision_context         TEXT,
     algorithm_ref            VARCHAR(500),
-    confidence_score         DOUBLE,
+    confidence_score         DOUBLE PRECISION,
     contestation_uri         VARCHAR(2000),
     human_override_available BOOLEAN,
     CONSTRAINT pk_ledger_supplement_compliance PRIMARY KEY (id),
